task2_animals/requirements.txt
kagglehub>=1.0.0
numpy>=2.2.6
pandas>=2.2.0
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
pytest>=8.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0
